requirements.txt
pandas==2.2.1
Pillow==10.3.0
plotly==5.20.0
pymongo==4.6.3
st_star_rating==0.0.6
streamlit==1.32.2
